+++ core/src/main/java/react4j/core/ReactErrorInfo.java
@@ -1,0 +1,13 @@
+package react4j.core;
+
+import javax.annotation.Nonnull;
+import jsinterop.annotations.JsPackage;
+import jsinterop.annotations.JsType;
+
+@JsType( isNative = true, namespace = JsPackage.GLOBAL, name = "Object" )
+public class ReactErrorInfo
+{
+  @SuppressWarnings( "NullableProblems" )
+  @Nonnull
+  public String componentStack;
+}