+++ processor/src/test/resources/input/com/example/lifecycle/OverrideLifecycleMethodsComponent.java
@@ -1,11 +1,13 @@
 package com.example.lifecycle;
 
+import elemental2.core.Error;
 import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
 import react4j.annotations.ReactComponent;
 import react4j.core.BaseProps;
 import react4j.core.BaseState;
 import react4j.core.Component;
+import react4j.core.ReactErrorInfo;
 import react4j.core.ReactNode;
 
 @ReactComponent
@@ -40,6 +42,11 @@ class OverrideLifecycleMethodsComponent
 
   @Override
   protected void componentWillUpdate( @Nullable final BaseProps nextProps, @Nullable final BaseState nextState )
+  {
+  }
+
+  @Override
+  protected void componentDidCatch( @Nonnull final Error error, @Nonnull final ReactErrorInfo info )
   {
   }