+++ CHANGELOG.md
@@ -37,6 +37,7 @@
 * 💥 **\[core\]** Remove support for string refs as they are deprecated in React js and slated for removal.
 * 💥 **\[core\]** Remove `ReactRefCallback` and replace it with `RefConsumer`.
 * 💥 **\[core\]** Remove `ReactElementChildren` class and replace usages with `ReactNode` class as they are equivalent.
+* 💥 **\[core\]** Add support for the `componentDidCatch` lifecycle method introduced in React 16.
 
 ### [v0.12](https://github.com/realityforge/react4j/tree/v0.12) (2017-11-09)
 [Full Changelog](https://github.com/realityforge/react4j/compare/v0.11...v0.12)